backend/requirements.txt
# ÉluIA Backend Dependencies - Production Pinned
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.29
pydantic==2.10.0
pydantic-settings==2.1.0
anthropic==0.18.1
pypdf2==3.0.1
python-docx==1.1.0
slowapi==0.1.9
aiofiles==23.2.1
sentence-transformers==2.5.1
numpy==1.26.4
scikit-learn==1.4.2
psycopg2-binary==2.9.9
alembic==1.13.1
